refactor(typedservice): move to dedicated folder
